pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atlas"
version = "0.1.0"
description = "Independent Research Project Environment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-language-conversations>=1.1.0",
    "azure-ai-projects>=1.0.0b11",
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.23.0",
    "azure-search-documents>=11.5.3",
    "azure-storage-blob>=12.25.1",
    "beautifulsoup4>=4.13.4",
    "chainlit>=2.6.2",
    "colorlog>=6.9.0",
    "fastmcp>=2.10.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.5",
    "mcp[cli]>=1.10.1",
    "numpy>=2.2.6",
    "ollama>=0.5.2",
    "openai>=1.84.0",
    "pandas>=2.3.2",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.5",
    "pyright>=1.1.403",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "semantic-kernel==1.32.2",
    "statsmodels>=0.14.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["E501"]
